requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
jinja2
httpx